[feat] masked softmax on npu2

+++ tests/dataflow/aie/external/exp.cc
@@ -18,8 +18,8 @@ void customized_exp(T_in *input_tensor, T_out *output_tensor) {
   const int F = HIDDEN / vec_factor;
   using vec_t = aie::vector<T_in, vec_factor>;
   event0();
-  aie::vector<bfloat16, vec_factor> log2e_vec =
-      aie::broadcast<bfloat16, vec_factor>(log2e);
+  aie::vector<float, vec_factor> log2e_vec =
+      aie::broadcast<float, vec_factor>(log2e);
   aie::vector<bfloat16, vec_factor> scale_vec =
       aie::broadcast<bfloat16, vec_factor>(1.0);
   for (int iter = 0; iter < SEQ_LEN; iter++) {
@@ -30,7 +30,8 @@ void customized_exp(T_in *input_tensor, T_out *output_tensor) {
       input_ptr += vec_factor;
       input_vec = aie::mul(input_vec, log2e_vec);
       auto result_vec = aie::exp2<bfloat16>(input_vec); // ! require XDNA2
-      aie::accum<accfloat, vec_factor> exp_out = aie::mul(result_vec, scale_vec);
+      aie::accum<accfloat, vec_factor> exp_out =
+          aie::mul(result_vec, scale_vec);
       aie::store_v(output_ptr, (exp_out.to_vector<float>()));
       output_ptr += vec_factor;
     }
